Delete framework stm32xxx_hal_conf.h when custom_config_header=yes
There is a bug where a custom `stm32xxx_hal_conf.h` header (enabled by
setting `build.stm32cube.custom_config_header=yes`) can't be used, as a
copy exists in the downloaded framework, and this copy is first on the
include path.

To address this, when `build.stm32cube.custom_config_header=yes`, delete
the framework copy of the header.